move MockErrorDownload
他のテストでも使うため
