+++ Tests/Runtime/DownloadAsyncDecorators/MockErrorDownload.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fbb63205b519849fe800288adafc0741
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
