+++ Tests/Runtime/Tasks/FetchBundlesTest.cs
@@ -44,34 +44,6 @@ namespace AssetBundleHubTests
             Assert.That(context.downloadedAssetBundles.Count, Is.EqualTo(0));
         });
 
-        public class MockErrorDownload : IDownloadAsyncDecorator<IDownloadRequestContext, IDownloadResponseContext>
-        {
-
-            int count = 0;
-            int throwCount;
-
-            /// <summary>
-            /// エラー出す用クラス
-            /// </summary>
-            /// <param name="throwCount"> 何番目のリクエストでエラーを出すか </param>
-            public MockErrorDownload(int throwCount = 1)
-            {
-                this.throwCount = throwCount;
-            }
-
-            public async UniTask<IDownloadResponseContext> DownloadAsync(IDownloadRequestContext context, CancellationToken cancellationToken, Func<IDownloadRequestContext, CancellationToken, UniTask<IDownloadResponseContext>> next)
-            {
-                cancellationToken.ThrowIfCancellationRequested();
-                count++;
-                if (count == throwCount)
-                {
-                    throw new Exception("test error: throw count");
-                }
-                await UniTask.DelayFrame(10);
-                return new DownloadResponseContext();
-            }
-        }
-
         [UnityTest]
         public IEnumerator 異常系_なんらかのエラーが起こったときに後続のダウンロード処理が中断されること() => UniTask.ToCoroutine(async () =>
         {
